# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AssessRequest": {
        "properties": {
          "lang": {
            "default": "zh",
            "description": "Output language for LLM-generated text: 'zh' or 'en'",
            "title": "Lang",
            "type": "string"
          },
          "query": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "AssessRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchRequest": {
        "properties": {
          "lang": {
            "default": "zh",
            "description": "Output language for LLM-generated text: 'zh' or 'en'",
            "title": "Lang",
            "type": "string"
          },
          "query": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "rewrite": {
            "default": false,
            "description": "Use LLM to rewrite query for better matching",
            "title": "Rewrite",
            "type": "boolean"
          },
          "top_k": {
            "default": 12,
            "maximum": 30.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Optional API key. Anonymous requests fall back to the free tier (60 req/min). Provide an X-API-Key header to access higher tiers.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Structural Team",
      "url": "https://structural.bytedance.city/"
    },
    "description": "跨领域结构同构搜索引擎 — Structural Isomorphism API.\n\nPublic surface for the Structural cross-domain knowledge engine. Endpoints are grouped by capability (Ask / Phase / Phenomenon / Pricing / Newsletter / Admin). Rate limits are tier-based; supply an `X-API-Key` header to promote a request beyond the anonymous free-tier bucket.",
    "license": {
      "name": "Proprietary"
    },
    "title": "Structural",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/search": {
      "post": {
        "operationId": "search_phenomena_api_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Phenomena",
        "tags": [
          "search"
        ]
      }
    },
    "/api/search/assess": {
      "post": {
        "description": "Run the LLM \"worthiness + rewrite\" pre-flight independently.\n\nThe frontend calls this in parallel with /api/search so it can show\nresults immediately and then overlay the coaching gate if the query\nscores below threshold. On any error we fail open (worth_score=5).",
        "operationId": "assess_query_api_search_assess_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assess Query",
        "tags": [
          "search"
        ]
      }
    },
...
  },
...
}